chore: move panel.js to ui/ subfolder
