+++ umap/static/umap/js/modules/global.js
@@ -2,7 +2,7 @@ import URLs from './urls.js'
 import Browser from './browser.js'
 import Facets from './facets.js'
 import Caption from './caption.js'
-import { Panel, EditPanel, FullPanel } from './panel.js'
+import { Panel, EditPanel, FullPanel } from './ui/panel.js'
 import Alert from './ui/alert.js'
 import Tooltip from './ui/tooltip.js'
 import * as Utils from './utils.js'
